examples/python/requirements.txt
opencv-python
Pillow
imageio